Untitled_folder/python3_requirements.txt
absl-py==0.7.0
alabaster==0.7.11
anaconda-client==1.7.2
anaconda-navigator==1.6.9
anaconda-project==0.8.2
apache-beam==2.11.0
appdirs==1.4.3
argcomplete==1.9.4
arrow==0.13.2
asn1crypto==0.24.0
astor==0.7.1
astroid==2.0.4
astropy==3.0.4
atomicwrites==1.2.1
attrs==18.2.0
Automat==0.7.0
autopep8==1.4.4
autosub==0.3.12
avro-python3==1.8.2
Babel==2.6.0
backcall==0.1.0
backports.shutil-get-terminal-size==1.0.0
beautifulsoup4==4.6.3
binaryornot==0.4.4
bitarray==0.8.3
bkcharts==0.2
blaze==0.11.3
bleach==2.1.4
bokeh==0.13.0
boto==2.49.0
boto3==1.9.98
botocore==1.12.98
Bottleneck==1.2.1
cachetools==3.1.0
certifi==2018.8.24
cffi==1.11.5
chardet==3.0.4
click==6.7
cloudpickle==0.5.5
clyent==1.2.2
colorama==0.3.9
conda==4.6.4
conda-build==3.0.27
conda-verify==2.0.0
configparser==3.7.1
constantly==15.1.0
contextlib2==0.5.5
cookiecutter==1.6.0
crcmod==1.7
cryptography==2.3.1
cycler==0.10.0
Cython==0.28.5
cytoolz==0.9.0.1
dask==0.19.1
databricks-cli==0.8.4
datashape==0.5.4
decorator==4.3.0
defusedxml==0.5.0
dill==0.2.9
distributed==1.23.1
dlib==19.17.0
docopt==0.6.2
docutils==0.14
entrypoints==0.2.3
enum34==1.1.6
et-xmlfile==1.0.1
fastavro==0.21.19
fastcache==1.0.2
fasteners==0.14.1
ffmpeg==1.4
filelock==3.0.8
Flask==0.12
Flask-Cors==3.0.6
future==0.17.1
gast==0.2.2
gevent==1.3.6
gitdb2==2.0.5
GitPython==2.1.11
glob2==0.6
gmpy2==2.0.8
google-api-core==1.8.1
google-api-python-client==1.7.8
google-apitools==0.5.26
google-auth==1.6.3
google-auth-httplib2==0.0.3
google-cloud-bigquery==1.6.1
google-cloud-bigtable==0.31.1
google-cloud-core==0.28.1
google-cloud-pubsub==0.39.0
google-resumable-media==0.3.2
googleapis-common-protos==1.5.8
greenlet==0.4.15
grpc-google-iam-v1==0.11.4
grpcio==1.18.0
gunicorn==19.9.0
h5py==2.8.0
hdfs==2.2.2
heapdict==1.0.0
html5lib==1.0.1
httplib2==0.11.3
humanize==0.5.1
hyperlink==18.0.0
idna==2.7
imageio==2.4.1
imageio-ffmpeg==0.3.0
imagesize==1.1.0
incremental==17.5.0
ipykernel==4.9.0
ipython==6.5.0
ipython-genutils==0.2.0
ipywidgets==7.4.1
isort==4.3.4
itsdangerous==0.24
jdcal==1.4
jedi==0.12.1
jeepney==0.3.1
Jinja2==2.10
jinja2-time==0.2.0
jmespath==0.9.3
joblib==0.13.2
jsonschema==2.6.0
jupyter==1.0.0
jupyter-client==5.2.3
jupyter-console==5.2.0
jupyter-core==4.4.0
jupyterlab==0.34.9
jupyterlab-launcher==0.13.1
Keras==2.2.0
Keras-Applications==1.0.2
Keras-Preprocessing==1.0.1
keyring==13.2.1
kitchen==1.2.5
kiwisolver==1.0.1
lazy-object-proxy==1.3.1
llvmlite==0.24.0
locket==0.2.0
lxml==4.2.5
Markdown==3.0.1
MarkupSafe==1.0
matplotlib==2.2.3
mccabe==0.6.1
mistune==0.8.3
mkl-fft==1.0.4
mkl-random==1.0.1
mleap==0.8.1
mlflow==0.8.2
mock==2.0.0
monotonic==1.5
more-itertools==4.3.0
mpmath==1.0.0
msgpack==0.5.6
multipledispatch==0.6.0
navigator-updater==0.1.0
nbconvert==5.4.0
nbformat==4.4.0
neptune-cli==2.8.23
networkx==2.1
nltk==3.3
nose==1.3.7
nose-exclude==0.5.0
notebook==5.6.0
numba==0.39.0
numexpr==2.6.8
numpy==1.15.4
numpydoc==0.8.0
nvidia-ml-py3==7.352.0
oauth2client==3.0.0
oauthlib==3.0.1
odo==0.5.1
olefile==0.46
opencv-python==4.0.0.21
openpyxl==2.5.6
packaging==17.1
pafy==0.5.4
pandas==0.23.4
pandas-ml==0.6.1
pandocfilters==1.4.2
parso==0.3.1
partd==0.3.8
path.py==11.1.0
pathlib2==2.3.0
patsy==0.5.0
pbr==5.1.3
pep8==1.7.1
pexpect==4.6.0
pickleshare==0.7.4
Pillow==5.2.0
pkginfo==1.4.2
pluggy==0.7.1
ply==3.11
poyo==0.4.2
proglog==0.1.9
progressbar==2.5
prometheus-client==0.3.1
prompt-toolkit==1.0.15
proto-google-cloud-datastore-v1==0.90.4
protobuf==3.7.0
psutil==5.4.7
ptyprocess==0.6.0
py==1.6.0
pyarrow==0.11.1
pyasn1==0.4.4
pyasn1-modules==0.2.2
pycodestyle==2.5.0
pycosat==0.6.3
pycparser==2.18
pycrypto==2.6.1
pycurl==7.43.0.2
pydot==1.2.4
pydot-ng==2.0.0
pydub==0.23.1
pyflakes==2.0.0
pygame==1.9.4
Pygments==2.2.0
PyJWT==1.7.1
pykwalify==1.5.2
pylint==2.1.1
pyodbc==4.0.24
pyOpenSSL==18.0.0
pyparsing==2.2.0
PySocks==1.6.8
PySoundFile==0.9.0.post1
pysrt==1.1.1
pytest==3.8.0
pytest-arraydiff==0.2
pytest-astropy==0.4.0
pytest-doctestplus==0.1.3
pytest-openfiles==0.3.0
pytest-remotedata==0.3.0
python-dateutil==2.7.3
pytube==9.5.1
pytz==2018.5
PyWavelets==1.0.0
PyYAML==3.13
pyzmq==17.1.2
QtAwesome==0.4.4
qtconsole==4.4.1
QtPy==1.5.0
querystring-parser==1.2.3
raven==6.10.0
readme-renderer==24.0
requests==2.19.1
requests-oauthlib==1.2.0
requests-toolbelt==0.9.1
resampy==0.2.1
rmate==1.0.3
rope==0.11.0
rsa==4.0
ruamel-yaml==0.15.46
s3transfer==0.2.0
scikit-image==0.14.0
scikit-learn==0.19.2
scipy==1.1.0
seaborn==0.9.0
SecretStorage==3.1.0
Send2Trash==1.5.0
service-identity==17.0.0
simplegeneric==0.8.1
simplejson==3.16.0
singledispatch==3.4.0.3
six==1.11.0
smmap2==2.0.5
snowballstemmer==1.2.1
sortedcollections==1.0.1
sortedcontainers==2.0.5
SoundFile==0.10.2
Sphinx==1.7.9
sphinxcontrib-websupport==1.1.0
spyder==3.3.1
spyder-kernels==0.2.6
SQLAlchemy==1.2.11
statsmodels==0.9.0
steppy==0.1.16
style==1.1.0
sympy==1.2
tables==3.5.1
tabulate==0.8.3
tblib==1.3.2
tensorboard==1.12.2
tensorboardX==1.6
tensorflow==1.12.0
tensorflow-data-validation==0.13.1
tensorflow-metadata==0.13.0
tensorflow-transform==0.13.0
termcolor==1.1.0
terminado==0.8.1
terminaltables==2.1.0
testpath==0.3.1
toolz==0.9.0
torch==1.0.1.post2
torchaudio==0.2
torchvision==0.2.2.post3
tornado==5.1
tqdm==4.26.0
traitlets==4.3.2
twine==1.13.0
Twisted==18.7.0
typed-ast==1.1.0
typing==3.6.6
unicodecsv==0.14.1
update==0.0.1
uritemplate==3.0.0
urllib3==1.23
voluptuous==0.11.5
wcwidth==0.1.7
webencodings==0.5.1
websocket-client==0.55.0
Werkzeug==0.14.1
whichcraft==0.5.2
widgetsnbextension==3.4.1
wrapt==1.10.11
xlrd==1.1.0
XlsxWriter==1.1.0
xlwt==1.3.0
youtube-dl==2019.3.9
zict==0.1.3
zipfile36==0.1.3
zope.interface==4.5.0
